t2/requirements.txt


aiohttp==3.8.5
asyncio==3.4.3
langchain==0.0.312
langchain-pinecone==0.0.1
llama-index==0.8.48
octoai==0.3.2
pinecone-client==2.2.4
pydantic==2.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2